32-bit console charcodes
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@7126 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ inc/fbcons.h
@@ -18,7 +18,7 @@ struct fbcons_seg {
     volatile uint32_t xpos;
     volatile uint32_t ypos;
 
-    volatile uint8_t  data[];
+    volatile uint32_t data[];
 };
 
 #endif
